Don't drop whitespace in lexer outer loop
The outer loop for lexing using the stream operator, which
acts as a formatted input stream. That means it strips whitespace
(including newlines).

We want to lex those ourselves, in the case of newline so that we
can increment the line counter.

Use `get` instead, which acts as an unformatted input operator so
leaves whitespace intact.