Use unique_ptr inside Expr variant
std::variant doesn't support incomplete types as template arugments
(because it needs to know how big they are). Since the types used
in the variant here are also reference the variant, there is no way
(that I can see) to define them all.

It so happens that we inted to use unique_ptr anyway (to avoid lots
of copying during tree construction, and to support possible tree
transformations later), and unique_ptr does support incomplete types
as template arguments.

So, use unique_ptr in the std::variant.