Rename JSON fields->children
More representative because this JSON element no longer really
represents the fields of the AST structs, because the fields
may now use unique_ptr in some places.

Children is also not perfect because the enums used by each struct
aren't really children of that struct, but at least it doesn't
imply that the text in the element is used directly to define
the structs' fields.