Extract some includes into JSON
Some includes are needed by the "framework" of the
code generator e.g. variant for making the "superclass".
But others e.g. string are used only by specific subclasses.

In the latter case these can be extracted into the JSON so that
they live alongside the data that uses them.