Mark ast factory funs as inline
They are global functions so they need to be marked
as inline since they are in a header file, which may
be included in multiple translation units in the same
compilation. That is an error (violates ODR) unless they
are marked as inline.

The definitions then need to be the same within each translation
unit but that should be the case because they come from the
same header file (unless something goes wrong...).