Add comment about how parser could be improved
I think it would reduce code repetition if these mapping
functions were extracted out into a single function
that does all the mapping for e.g. Token->BinOp, and throws
for tokens that it can't convert. That would move the
conversion logic off to the side, so it doesn't clutter
up the parsing.