Consume token expected by `expect`
`expect` is supposed to assert that a given token
is coming next but it should also then consume that token.

Currently it doesn't, which is causing a bug when parsing groupings.
The parser expects a right bracket after the grouping but doesn't
consume it, which messes up the rest of the parsing sequence.