Prepend START token to lexer output, use it in parser
The parser is currently broken because it tries to set a size_t
to -1 but size_t is unsigned so that doesn't work.

Using a signed long and implicitly converting doesn't play well
and making the column number optional before any tokens have
been consumed would require a lot of refactoring and adding
in extra code paths.

An easier solution is to add a START token, similar to EOF, at the
start of every lex, which the parser expects. The token doesn't provide
any useful information but it makes the parser simpler because there is
no special case for when no tokens have been consumed yet. (Again,
similar to the motivation for adding an explicit EOF.)